undo refactor of moving proc-macros into separate crate
