+++ Cargo.toml
@@ -11,8 +11,16 @@ readme = "README.md"
 keywords = ["builder"]
 categories = ["rust-patterns"]
 
+[lib]
+proc-macro = true
+
 [dependencies]
-typed-builder-proc-macros.path = "proc-macros"
+either = "1"
+prettyplease = { version = "0.1", optional = true }
+proc-macro2 = "1"
+quote = "1"
+regex = "1"
+syn = { version = "1", features = ["full", "extra-traits"] }
 
 [features]
-debug = ["typed-builder-proc-macros/debug"]
+debug = ["dep:prettyplease"]
