--- proc-macros/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "typed-builder-proc-macros"
-description = "Compile-time type-checked builder derive"
-version = "0.11.0"
-authors = ["IdanArye <idanarye@gmail.com>", "Chris Morgan <me@chrismorgan.info>"]
-edition = "2018"
-license = "MIT/Apache-2.0"
-repository = "https://github.com/idanarye/rust-typed-builder"
-documentation = "https://idanarye.github.io/rust-typed-builder/"
-readme = "README.md"
-keywords = ["builder"]
-categories = ["rust-patterns"]
-
-[lib]
-proc-macro = true
-
-[dependencies]
-either = "1"
-prettyplease = { version = "0.1", optional = true }
-proc-macro2 = "1"
-proc-macro-crate = "1"
-quote = "1"
-regex = "1"
-syn = { version = "1", features = ["full", "extra-traits"] }
-
-[features]
-debug = ["dep:prettyplease"]
